int->size_t

+++ examples/cpp/routing-get-clusters/s4u-routing-get-clusters.cpp
@@ -30,9 +30,9 @@ int main(int argc, char* argv[])
 
   for (auto d : dragonfly_clusters) {
     XBT_INFO("%s' dragonfly topology:", d->get_cname());
-    for (int i = 0; i < d->get_host_count(); i++) {
+    for (size_t i = 0; i < d->get_host_count(); i++) {
       const simgrid::kernel::routing::DragonflyZone::Coords coords = d->rankId_to_coords(i);
-      XBT_INFO("   %d: (%lu, %lu, %lu, %lu)", i, coords.group, coords.chassis, coords.blade, coords.node);
+      XBT_INFO("   %zu: (%lu, %lu, %lu, %lu)", i, coords.group, coords.chassis, coords.blade, coords.node);
     }
   }
